[FIX] fix render bug

+++ include/Api.hpp
@@ -2,6 +2,9 @@
 #define __API_HPP__
 
 #include <string>
+#include <SDL.h>
+#include <SDL_image.h>
+
 /* ### Singleton class ### */
 class Object;
 class Api{
@@ -16,8 +19,8 @@ public:
   static void destroy();
   void LoadGraphic(Object* target,std::string filename);
   void SetTexture(Object* target,std::string filename);
-  void SetGraphicRect(Object* target,GLfloat sx,GLfloat sy,GLfloat dx,GLfloat dy);
-  void DrawGraphic(Object* target,GLfloat x,GLfloat y);
+  void SetGraphicRect(Object* target,Uint32 sx,Uint32 sy,Uint32 dx,Uint32 dy);
+  void DrawGraphic(Object* target,Uint32 x,Uint32 y);
 };
 
 #endif
